+++ global/scripts/languages/golang/cyclonedx/run.sh
@@ -21,11 +21,18 @@ if [ -d "pkg" ]; then
   echo "Found 'pkg' directory, using 'cyclonedx-gomod mod' command..."
   "$(go env GOPATH)/bin/cyclonedx-gomod" mod -json -output "$BOM_PATH/bom.json" -licenses
 else
-  folder="$(find . -type f -name main.go -not -path '*/.go/*' -exec dirname {} \;)"
-  if [ -z "$folder" ]; then
+  folders="$(find . -type f -name main.go -not -path '*/.go/*' -exec dirname {} \;)"
+  if [ -z "$folders" ]; then
     echo "Could not find a directory containing Go files"
     exit 1
   fi
-  echo "Using 'cyclonedx-gomod app' command..."
-  "$(go env GOPATH)/bin/cyclonedx-gomod" app -json -output "$BOM_PATH/bom.json" -packages -files -licenses -main "$folder"
+
+  folder_count=$(echo "$folders" | wc -l)
+  if [ "$folder_count" -eq 1 ]; then
+    echo "Using 'cyclonedx-gomod app' command..."
+    "$(go env GOPATH)/bin/cyclonedx-gomod" app -json -output "$BOM_PATH/bom.json" -packages -files -licenses -main "$folders"
+  else
+    echo "Found $folder_count main packages, using 'cyclonedx-gomod mod' command..."
+    "$(go env GOPATH)/bin/cyclonedx-gomod" mod -json -output "$BOM_PATH/bom.json" -licenses
+  fi
 fi